+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Added index file support for Flex Objects
 1. [](#improved)
     * Improved error detection for broken Flex Objects
+    * Removed apc and xcache support, made apc alias of apcu
 
 # v1.6.0-rc.1
 ## 01/30/2019
@@ -12,14 +13,12 @@
 1. [](#improved)
     * Improved `$page->forms()` call, added `$page->addForms()`
     * Made `FormFlashFile` more robust against deleted files
-    * Removed apc and xcache support, made apc alias of apcu
     * Updated languages from crowdin
 1. [](#bugfix)
     * Fixed a bug in `FormFlashFile::moveTo()` not deleting the old file
     * Fixed `FlexMediaTrait::getMedia()` trying to include uploaded but already moved media
     * Fixed `ImageMedium` constructor warning when file does not exist
     * Fixed bad host header in PSR-7 (if using 'php -S localhost:8000 system/router.php')
-    * Fixed Getters::count() method not returning anything
     * Fixed non-namespaced exceptions in scheduler
     
 # v1.6.0-beta.8